tic608_nsst_rt: While running the regression tests on Gyre, it was noted that the new global_4denvar_T126 regression test was failing due to non-reproducible results.  Looking at the stdout files for the loproc_updat and loproc_contrl runs, the loproc_updat stdout had instances of:
Doing nothing in dummy routine NST_INIT_

and similar messages, while the loproc_contrl did not have these.  This message is found in stub_nstmod.f90, which shouldn't be compiled.  To correct this issue, I removed stub_nstmod.f90 from the rest of the source files in src/CMakeLists.txt.

I also modified the resources to run the new script for WCOSS in regression_param.sh.