Allocate fhour in write_sfc_ prior to use (#483) (#484)
This PR is opened to request merger of a bug fix for `write_sfc_` (file
`src/gsi/netcdfgfs_io.f90`). The version of `write_sfc_` in the
authoritative repository does not allocate local array `fhour` prior to
use. This bug causes `gsi.x` to seg fault when writing the netcdf
`sfcanl_gsi` output file. The fix in this PR is simple: allocate `fhour`
prior to use.

Merger of this PR into the authoritative develop will close issue #483.