Feature/enkf q2 (#568)
Resubmit PR for cleaning the unneeded item in the previous PR 
In correspondence to the EMC GSI Issue#566, this PR contains a quick
adding of the clipping of negative values of sphum (q) in the analysis
of FV3-LAM EnKF.
This part of codes are not tested in the current GSI regression tests,
which, hence, are not run.
The current codes are verified using local FV3-LAM case. It is found the
differences from this changes exist for sphum ( maximum values about
0.003 (units) and 0.3 K for T (sensible T). The latter is because the
sphum would be used when the analysis TV is converted to T. All
differences are on spontaneous points and values are reasonable as
expected.
Hence, the code is regarded as verified.

Fixes #566

**DUE DATE for this PR is 6/15/2023.** If this PR is not merged into
`develop` by this date, the PR will be closed and returned to the
developer.