Gsi fed (#590)
<!-- PLEASE READ -->
<!--
Before opening a PR, please note these guidelines:

- Each PR should only address ONE topic and have an associated issue
- No hardcoded or paths to personal directories should be present
- No temporary or backup files should be committed
- Any code that was disabled by being commented out should be removed
-->

**Description**

Initialization of the operational RRFSv1 will include assimilation of
flash-extent density (FED) observations from the GOES Geostationary
Lightning Mapper (GLM). The current PR is the first of at least 3 that
will be needed to introduce the capability of FED assimilation into the
code and regional workflow. The new capabilities that are added to GSI
are:

* reading NetCDF FED observations
* applying an observation operator that maps the model state to FED.

Much of the code was originally developed by Rong Kong at OU-CAPS (Kong
et al. 2020, Wang et al. 2021, Kong et al. 2022;
https://doi.org/10.1175/MWR-D-19-0192.1,
https://doi.org/10.1175/MWR-D-20-0406.1,
https://doi.org/10.1175/MWR-D-21-0326.1). Recently, the observation
operator has been modified by Amanda Back and Ashley Sebok based on
tests with regional, convection-allowing FV3 forecasts. The new
observation operator includes a cap of 8 flashes / minute for both the
observed and simulated FED.

The observation operator is specific to the 3-km regional FV3
application in RRFS. Development of a more general observation operator
is left to future work.

Fixes #588 

<!-- Please include relevant motivation and context. -->
<!-- Please include a summary of the change and which issue is fixed.
-->
<!-- List any dependencies that are required for this change. -->

<!-- Please provide reference to the issue this pull request is
addressing. -->
<!-- For e.g. Fixes #IssueNumber -->

**Type of change**

Please delete options that are not relevant.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

**How Has This Been Tested?**

Initial tests were with NOAA-EMC GSI-EnKF code obtained in April 2023
and modified to include the assimilation of FED observations. A
prototype of RRFSv1 was cycled hourly for 2.5 days, and the EnKF
assimilation included FED data assimilation.

For the current PR, only the GSI observer with FED (and radar
reflectivity) observations was tested. It produces identical results to
those obtained in April 2023.

<!-- Please describe the tests that you ran to verify your changes and
on the platforms these tests were conducted. -->
<!-- Provide instructions so we can reproduce. -->
<!-- Please also list any relevant details for your test configuration
-->
  
**Checklist**

- [ ] My code follows the style guidelines of this project
- [X] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] New and existing tests pass with my changes
- [ ] Any dependent changes have been merged and published

**DUE DATE for this PR is 8/24/2023.** If this PR is not merged into
`develop` by this date, the PR will be closed and returned to the
developer.

---------

Co-authored-by: Ming Hu <Ming.Hu@noaa.gov>