Refine PM2.5 DA for the RRFS_SD model  (#609)
**Description**

Refine the PM2.5 DA for the RRFS_SD model by use of veg_type, which is
used to decide whether the obs is in urban area or not. Different
thresholds for innovations outside/inside urban areas will be used. Add
new namelist parameters, such as threshold for innovations, anowbufr
type
Read in station terrain height, PM10 et al if extended BUFR format for
anow air quality data is used

Fixes #606 


**Type of change**

Please delete options that are not relevant.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

**How Has This Been Tested?**

<!-- This change will only affect PM2.5 DA for RRFS_SD, so it will not
affect any other tests -->

  
**Checklist**

- [X ] My code follows the style guidelines of this project
- [X] I have performed a self-review of my own code
- [X] I have commented my code, particularly in hard-to-understand areas
- [ ] New and existing tests pass with my changes
- [ ] Any dependent changes have been merged and published

**DUE DATE for this PR is 9/22/2023.** If this PR is not merged into
develop by this date, the PR will be closed and returned to the
developer.