Bug fix of GPS (GNSS)-RO assimilation with multiple bufr data files
The RO assimilation needs to assign a unique ID for each profile, used for superobing
at vertical levels (inflation of observation errors).
In the original code, when multiple bufr data files are inputted, duplicated
profile IDs can occur and lead to an unexpected result of observation error inflation.
The revised code re-computes the RO profile ID in the 'obs_para' subroutine
to ensure its uniqueness and thus fix the problem.