Two small tweaks to thinning (#734)
This PR addresses GSI #731. It modifies the satthin.F90 routine to allow
data to be processed unthinned in the GSI without requiring exorbitant
resources and it removes an unnecessary thinning from the read_atms.f90
routine.

The change to satthin reduces the value of `itxmax` from 1.e9 to 1.e7
which should be sufficient for most sensors. But if a larger number is
required for a sensor, this may be passed in via optional argument.

This change will change results for ATMS only as documented in the
issue.