update gsi code and convinfo for saildrone observations (#795)
We need to be able to switch of active assimilation of saildrone data in
operational GSI before obsproc1.3 is implemented. This requires a one
line code change.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update

The change has been tested in a single-cycle GSI run. The saildrone
observations are given the subtype 02 and are not actively assimilated.
  
**Checklist**

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] New and existing tests pass with my changes
- [x] Any dependent changes have been merged and published