Update `fix/global_convinfo.txt` to allow assimilation of GOES-19 AMVs in release/gfsda.v16 (#842)
Update `fix/global_convinfo.txt` to allow assimilation of GOES-19 AMVs.

According to [this
site](https://codes.ecmwf.int/odb/satelliteidentifier/), the BUFR ID for
GOES-19 is 273.

After merging this needs to be tagged as `gfsda.v16.3.23`

**Type of change**

Please delete options that are not relevant.
- [x] New feature (non-breaking change which adds functionality)


**How Has This Been Tested?**

Checked out in `global-workflow` to ensure fix file is as expected.
  
**Checklist**

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] New and existing tests pass with my changes
- [x] Any dependent changes have been merged and published